% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvDat.R
\name{convDat}
\alias{convDat}
\title{convDat Funktion}
\usage{
convDat(inputDF, searchType, targetType)
}
\arguments{
\item{inputDF}{A data.frame}

\item{searchType}{Type of the columns which have to be converted.}

\item{targetType}{Type to which the columns with type "searchType" will be converted}
}
\value{
a data.frame with column types converted
}
\description{
Converts all columns of a data frame with type X to type Y.
}
\examples{
\dontrun{convDat(iris, is.factor, as.character)}
}
