% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacknife.R
\name{gamma_est}
\alias{gamma_est}
\title{gamma_jacknife Funktion}
\usage{
gamma_est(data)
}
\arguments{
\item{data}{A numeric vector}
}
\value{
a numeric vector containing the parameter estimates
}
\description{
Computes the jacknife parameter estimates of the gamma distribution
}
\examples{
\dontrun{gamma_jackknife(cats$Hwt)}
}
